[refactor]:Modified the file structure in channel module.
