+++ issue_tracker/channels/backends/discord_backend.py
@@ -3,34 +3,14 @@ import json
 from discord_webhook import DiscordWebhook, DiscordEmbed
 from django.core.exceptions import ImproperlyConfigured
 
-
-class Channel:
-    """
-    Base class for issue tracker channels.
-    """
-    def __init__(self, name: str):
-        """
-        Initializes Channel object.
-
-        Args:
-            name (str): The name of the issue tracker channel.
-        """
-        self.name = name
-
-    def send_notification(self, **kwargs):
-        """
-        Sends a notification to the issue tracker channel.
-
-        Raises:
-            NotImplementedError: If the send_notification method is not implemented in a subclass.
-        """
-        raise NotImplementedError("send_notification method not implemented in the base class")
+from issue_tracker.channels.channel import Channel
 
 
 class DiscordChannel(Channel):
     """
     Sends notifications to a Discord channel.
     """
+
     def send_notification(self, **kwargs):
         """
         Sends a notification to a Discord channel.
